+++ android/src/main/java/com/gokberkbar/locale_plus/MethodNames.java
@@ -12,7 +12,8 @@ public enum MethodNames {
     getPmSymbol("getPmSymbol"),
     getTimeZoneIdentifier("getTimeZoneIdentifier"),
     isUsingSamsungKeyboard("isUsingSamsungKeyboard"),
-    getFirstDayOfWeek("getFirstDayOfWeek");
+    getFirstDayOfWeek("getFirstDayOfWeek"),
+    getDateFormatPattern("getDateFormatPattern");
 
     private final String text;